add changes after review
